// src/types/openai-types.ts
export namespace openai {
	export interface CreateChatCompletionDeltaResponse {
		id: string
		object: 'chat.completion.chunk'
		created: number
		model: string
		choices: [
			{
				delta: {
					role: Role
					content?: string
				}
				index: number
				finish_reason: string | null
			}
		]
	}

	/**
	 *
	 * @export
	 * @interface ChatCompletionRequestMessage
	 */
	export interface ChatCompletionRequestMessage {
		/**
		 * The role of the author of this message.
		 * @type {string}
		 * @memberof ChatCompletionRequestMessage
		 */
		role: ChatCompletionRequestMessageRoleEnum
		/**
		 * The contents of the message
		 * @type {string}
		 * @memberof ChatCompletionRequestMessage
		 */
		content: string | [{}]
		/**
		 * The name of the user in a multi-user chat
		 * @type {string}
		 * @memberof ChatCompletionRequestMessage
		 */
		name?: string
	}
	export declare const ChatCompletionRequestMessageRoleEnum: {
		readonly System: 'system'
		readonly User: 'user'
		readonly Assistant: 'assistant'
	}
	export declare type ChatCompletionRequestMessageRoleEnum =
		(typeof ChatCompletionRequestMessageRoleEnum)[keyof typeof ChatCompletionRequestMessageRoleEnum]
// ...
	export interface CreateChatCompletionRequest {
// ...
		messages: Array<ChatCompletionRequestMessage>
// ...
	}
// ...
}